requirements.txt
Flask
mysql-connector-python
bcrypt
twilio
pyttsx3
